gtsrb-pytorch/requirements.txt
torchvision~=0.15.1+cu117
numpy~=1.24.2
Pillow~=9.4.0
pandas~=1.5.3
tqdm~=4.65.0
torcheval~=0.0.6
seaborn~=0.12.2
matplotlib~=3.7.1
tabulate~=0.9.0